+++ src/components/list.js
@@ -1,10 +1,14 @@
-import React from "react";
+import React, { useState } from "react";
 import { Button, ButtonGroup, Grid } from "@material-ui/core";
 import { green } from "../ui/colors";
 import ProductCard from "../components/card";
 
-export default function Lista(props) {
-  const product = props.product.map((product, index) => (
+export default function Lista({ products }) {
+  const itemsPerPage = 9;
+  const [items, setItems] = useState([...products].splice(0, itemsPerPage));
+  const [currentPage, setCurrentPage] = useState(0);
+
+  const product = products.map((product, index) => (
     <ProductCard
       key={index}
       name={product.nombre}
@@ -13,9 +17,39 @@ export default function Lista(props) {
     />
   ));
 
+  //BOTON SIGUIENTE
+  const nextHandler = () => {
+    const totalItems = products.length;
+
+    const nextPage = currentPage + 1;
+
+    const firstIndex = nextPage * itemsPerPage;
+
+    if (firstIndex === totalItems) return;
+
+    setItems([...products].splice(firstIndex, itemsPerPage));
+    setCurrentPage(nextPage);
+
+    window.scrollTo({ top: 300, behavior: "smooth" });
+  };
+
+  //BOTON ANTERIOR
+  const prevHandler = () => {
+    const prevPage = currentPage - 1;
+
+    if (prevPage < 0) return;
+
+    const firstIndex = prevPage * itemsPerPage;
+
+    setItems([...products].splice(firstIndex, itemsPerPage));
+    setCurrentPage(prevPage);
+
+    window.scrollTo({ top: 300, behavior: "smooth" });
+  };
+
   return (
     <>
-      <Grid style={{ minHeight: 2300 }}>
+      <Grid style={{ minHeight: 1500 }}>
         <Grid
           style={{ padding: 30 }}
           container
@@ -27,7 +61,7 @@ export default function Lista(props) {
       </Grid>
       <ButtonGroup style={{ marginBottom: 40 }} orientation="horizontal">
         <Button
-          onClick={props.prevHandler}
+          onClick={prevHandler}
           style={{
             backgroundColor: "#fed227",
             color: `${green}`,
@@ -39,7 +73,7 @@ export default function Lista(props) {
           Anterior
         </Button>
         <Button
-          onClick={props.nextHandler}
+          onClick={nextHandler}
           style={{
             backgroundColor: "#fed227",
             color: `${green}`,
